pyproject.toml
[project]
name = "a2a-demo"
version = "0.1.0"
description = "A2A Protocol Demo - LangGraph Agents with MCP Tools"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
a2a-research = "a2a_demo.agents.research:main"
a2a-writer = "a2a_demo.agents.writer:main"
a2a-client = "a2a_demo.client:main"
